+++ tests/test_app.py
@@ -1,12 +1,15 @@
+import pytest
 from flask import Blueprint
 from flask.views import MethodView
 from openapi_spec_validator import validate_spec
 
+from .schemas import BarSchema
 from .schemas import FooSchema
 from .schemas import PaginationSchema
 from apiflask import APIBlueprint
 from apiflask import APIFlask
 from apiflask import input
+from apiflask import output
 from apiflask import Schema
 from apiflask.fields import Integer
 from apiflask.fields import String
@@ -168,3 +171,49 @@ def test_dispatch_static_request(app, client):
     app.register_blueprint(bp, url_prefix='/foo')
     rv = client.get('/foo/static/hello.css')  # endpoint: foo.static
     assert rv.status_code == 404
+
+
+def schema_name_resolver1(schema):
+    name = schema.__class__.__name__
+    if schema.partial:
+        name += '_'
+    return name
+
+
+def schema_name_resolver2(schema):
+    name = schema.__class__.__name__
+    if name.endswith('Schema'):
+        name = name[:-6] or name
+    if schema.partial:
+        name += 'Partial'
+    return name
+
+
+@pytest.mark.parametrize('resolver', [
+    APIFlask._schema_name_resolver,
+    schema_name_resolver1,
+    schema_name_resolver2
+])
+def test_schema_name_resolver(app, client, resolver):
+    app.schema_name_resolver = resolver
+
+    @app.route('/foo')
+    @output(FooSchema)
+    def foo():
+        pass
+
+    @app.route('/bar')
+    @output(BarSchema(partial=True))
+    def bar():
+        pass
+
+    spec = app.spec
+    if resolver == schema_name_resolver1:
+        assert 'FooSchema' in spec['components']['schemas']
+        assert 'BarSchema_' in spec['components']['schemas']
+    elif resolver == schema_name_resolver2:
+        assert 'Foo' in spec['components']['schemas']
+        assert 'BarPartial' in spec['components']['schemas']
+    else:
+        assert 'Foo' in spec['components']['schemas']
+        assert 'BarUpdate' in spec['components']['schemas']